refactor: db types moved to its own file
