+++ src/components/SearchBox/SearchBox.tsx
@@ -1,8 +1,8 @@
 import { useNotesContext } from '@/context/NotesContext';
-import { Note } from '@/database/db';
+import type { Note } from '@/types/dbTypes';
 import { AppRoutes } from '@/types/generalTypes';
 import { Button, rem } from '@mantine/core';
-import { Spotlight, spotlight, SpotlightActionData } from '@mantine/spotlight';
+import { Spotlight, spotlight } from '@mantine/spotlight';
 import { IconFileText, IconSearch } from '@tabler/icons-react';
 import { useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
