+++ src/types/dbTypes.ts
@@ -1,3 +1,15 @@
-export type RegisteredUsersT = {
-  [nickname: string]: string;
-};
+interface User {
+  id: number;
+  nickname: string;
+  hashedPass: string;
+  signedIn: number;
+}
+
+interface Note {
+  id: number;
+  userId: number;
+  title: string;
+  content: string;
+}
+
+export type { User, Note };
